indexing synapse rois for faster computation of roi counts; computing meta roi info and pre/post counts after synapse load; segment addition and connection info addition are separate methods

+++ neuprint-load-procedures/src/main/java/org/janelia/flyem/neuprintloadprocedures/model/RoiInfo.java
@@ -58,7 +58,7 @@ public class RoiInfo {
      * @param pre  presynaptic density count
      * @param post postsynaptic density count
      */
-    public void addSynapseCountsForRoi(String roi, int pre, int post) {
+    public void addSynapseCountsForRoi(String roi, long pre, long post) {
         this.synapseCountsPerRoi.put(roi, new SynapseCounter(pre, post));
     }
